+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and `std:num:int_to_open_closed01`.
 * Feature: Added `std:rand:split_mix64_next_closed_open01` and
 `std:rand:split_mix64_next_open_closed01`.
 * Bugfix: Protect against division by 0 and make a wlambda panic from it.
+* Feature: Added basic prototyped object orientation.
 
 0.3.6 (2020-01-24)
 ==================